+++ src/LDL/Framework/Base/Collection/Key/Resolver/Contracts/FullKeyResolverInterface.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace LDL\Framework\Base\Collection\Key\Resolver\Contracts;
-
-interface FullKeyResolverInterface extends CustomKeyResolverInterface, DuplicateKeyResolverInterface, NullKeyResolverInterface
-{
-
-}